feat-177: editor formatter and calculate volume
